pyproject.toml
[tool.poetry]
name = "airflow"
version = "0.1.0"
description = "A pip manager for all packages"
authors = ["Adrien Crapart <adrien.crapart@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "3.9.2"
apache-airflow = "2.7.0"
pandas = "2.1.1"
pymysql = "1.1.0"
docker = "6.1.3"
selenium = "4.15.2"
beautifulsoup4 = "4.12.2"
pymongo = "4.5.0"
requests = "2.31.0"
swifter = "1.4.0"
numpy = "^1.26.4"
lxml = "^5.2.2"
flask-bcrypt = "^1.0.1"
boto3 = "^1.34.119"
openpyxl = "^3.1.3"
pendulum = "^3.0.0"
retrying = "^1.3.4"
gitpython = "^3.1.43"
pybitbucket37 = "^0.0.2"
geopandas = "^0.14.4"
feedparser = "^6.0.11"
rtree = "^1.2.0"
chardet = "^5.2.0"
geoalchemy2 = "^0.15.1"
tabulate = "^0.9.0"
bleach = "^6.1.0"
python-dotenv = "^1.0.1"
tqdm = "^4.66.4"
pyarrow = "^16.1.0"
py7zr = "^0.21.0"
requests-toolbelt = "^1.0.0"
asgiref = "^3.8.1"
setuptools = "^70.0.0"
sqlalchemy = "<=2.0"
shapely = "^2.0.4"
google-cloud-aiplatform = "^1.55.0"
apache-airflow-providers-amazon = "^8.23.0"
apache-airflow-providers-slack = "^8.7.1"
apache-airflow-providers-ssh = "^3.11.1"
apache-airflow-providers-ftp = "^3.9.1"
apache-airflow-providers-mongo = "^4.1.1"
apache-airflow-providers-jenkins = "^3.6.1"
apache-airflow-providers-postgres = "^5.11.1"
apache-airflow-providers-mysql = "^5.6.1"
apache-airflow-providers-docker = "^3.12.0"
apache-airflow-providers-samba = "^4.7.1"
apache-airflow-providers-odbc = "^4.6.1"
apache-airflow-providers-github = "^2.6.1"
apache-airflow-providers-http = "^4.11.1"
apache-airflow-providers-redis = "^3.7.1"
apache-airflow-providers-smtp = "^1.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
